GH-14: Move ValidTokenSchema to utils
